calendar: google: fetch events until end of week in current month
Previous code would fetch events from the first day of the week, of the
first week in the current month, and 31 days forward. This change uses
the NextMonth() helper to set the end date to the first day of the week
of the first week in the next month. Phew!